// src/app/api/models/medical-procedure.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Detalles del Medical Procedure a crear.
 */
export interface MedicalProcedure {

  /**
   * Status of the procedure
   */
  active?: boolean;

  /**
   * Procedure Code (CUPS, CPT, etc.)
   */
  code: string;

  /**
   * Identificador único
   */
  id?: number;

  /**
   * Procedure Name
   */
  name: string;

  /**
   * Type of procedure
   */
  type?: string;
}